pyproject.toml
[build-system]
requires = ['setuptools >= 61.0.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'whippersnappy'
version = '2.0.0-dev'
description = 'A package to plot and capture FastSurfer and FreeSurfer-style surface overlays.'
readme = 'README.md'
license = {file = 'LICENSE'}
requires-python = '>=3.9'
authors = [
    {name = 'Martin Reuter', email = 'martin.reuter@dzne.de'},
]
maintainers = [
    {name = 'Martin Reuter', email = 'martin.reuter@dzne.de'},
]
keywords = [
    'python',
    'FreeSurfer',
    'Brain MRI',
    'Cortical Surface',
]
classifiers = [
    'Operating System :: Microsoft :: Windows',
    'Operating System :: Unix',
    'Operating System :: MacOS',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Natural Language :: English',
    'License :: OSI Approved :: MIT License',
    'Intended Audience :: Science/Research',
]
dependencies = [
    'glfw',
    'numpy>=1.21',
    'pyrr',
    'pillow',
    'pyopengl==3.1.6',
    'nibabel',
    'PyQt6',
    'psutil'
]

[project.optional-dependencies]
build = [
    'build',
    'twine',
]
doc = [
    'furo!=2023.8.17',
    'matplotlib',
    'memory-profiler',
    'numpydoc',
    'sphinx!=7.2.*',
    'sphinxcontrib-bibtex',
    'sphinx-copybutton',
    'sphinx-design',
    'sphinx-gallery',
    'sphinx-issues',
    'pypandoc',
    'nbsphinx',
    'IPython', # For syntax highlighting in notebooks
    'ipykernel',
]
notebook = [
    'pythreejs',    # Three.js for interactive 3D (works in all Jupyter environments)
    'ipywidgets',   # Required for pythreejs
]
style = [
    'bibclean',
    'codespell',
    'pydocstyle[toml]',
    'ruff',
]
test = [
    'pytest',
    'pytest-cov',
    'pytest-timeout',
]
all = [
    'whippersnappy[build]',
    'whippersnappy[doc]',
    'whippersnappy[style]',
    'whippersnappy[test]',
    'whippersnappy[notebook]',
]
full = [
    'whippersnappy[all]',
]

[project.urls]
homepage = 'https://github.com/Deep-MI/WhipperSnapPy'
documentation = 'https://github.com/Deep-MI/WhipperSnapPy'
source = 'https://github.com/Deep-MI/WhipperSnapPy'
tracker = 'https://github.com/Deep-MI/WhipperSnapPy/issues'

[project.scripts]
whippersnap = 'whippersnappy.cli:run'
whippersnappy-sys_info = 'whippersnappy.commands.sys_info:run'

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ['whippersnappy*']
exclude = ['whippersnappy*tests']

[tool.setuptools.package-data]
whippersnappy = ['resources/fonts/*.ttf']

[tool.pydocstyle]
convention = 'numpy'
ignore-decorators = '(copy_doc|property|.*setter|.*getter|pyqtSlot|Slot)'
match = '^(?!setup|__init__|test_).*\.py'
match-dir = '^whippersnappy.*'
add_ignore = 'D100,D104,D107'

[tool.ruff]
line-length = 120
extend-exclude = [
    "doc",
    "setup.py",
]

[tool.ruff.lint]
# https://docs.astral.sh/ruff/linter/#rule-selection
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "I",   # isort
    # "SIM", # flake8-simplify
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = '6.0'
addopts = '--durations 20 --junit-xml=junit-results.xml --verbose'
filterwarnings = []

[tool.coverage.run]
branch = true
cover_pylib = false
omit = [
    '**/__init__.py',
    '**/whippersnappy/_version.py',
    '**/whippersnappy/commands/*',
    '**/tests/**',
]

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'if __name__ == .__main__.:',
]
precision = 2
